Moving google calender events to qrcode page
