+++ app/controllers/users_controller.rb
@@ -1,3 +1,6 @@
+require 'google/api_client/client_secrets.rb'
+require 'google/apis/calendar_v3'
+
 class UsersController < ApplicationController
 
   before_action :logged_in?, only: [:index, :show, :edit, :update]
@@ -9,6 +12,7 @@ class UsersController < ApplicationController
   def index
     @user = current_user
     @qr = encode
+    @events = get_events
   end
 
   def show
@@ -54,6 +58,52 @@ class UsersController < ApplicationController
 
   def encode
     RQRCode::QRCode.new(session[:user_id].to_s)
+  end
+
+  def get_events
+    # Initialize Google Calendar API
+    service = Google::Apis::CalendarV3::CalendarService.new
+    # Use google keys to authorize
+    service.authorization = google_secret.to_authorization
+    # Request for a new aceess token just incase it expired
+    service.authorization.refresh!
+    # Get a list of events
+    calendar_id = 'primary'
+    start_date = Date.today.rfc3339
+    end_date = (Date.today + 1).rfc3339
+    response = service.list_events(calendar_id,
+                                single_events: true,
+                                order_by: 'startTime',
+                                time_min: start_date,
+                                time_max: end_date
+                              )
+    # @events << 'No events found' if response.items.empty?
+    @events = []
+    response.items.each do |event|
+      if event.attendees != nil
+        event.attendees.each do |attendee|
+          if attendee.email == current_user.email && attendee.response_status != "declined"
+            @events << "#{event.summary} (#{Date.today.strftime("%d.%m.%Y")})"
+          end
+        end
+      else 
+        @events << "#{event.summary} (#{Date.today.strftime("%d.%m.%Y")})"
+      end
+    end
+
+    @events
+  end
+  
+  def google_secret
+    Google::APIClient::ClientSecrets.new(
+      { "web" =>
+        { "access_token" => current_user.google_token,
+          "refresh_token" => current_user.google_refresh_token,
+          "client_id" => ENV['GOOGLE_CLIENT_ID'],
+          "client_secret" => ENV['GOOGLE_CLIENT_SECRET']
+        }
+      }
+    )
   end
 
 end
