+++ app/views/users/index.html.erb
@@ -17,9 +17,15 @@
          <% end -%>
         </table>
       </p>
+
       <div class="row align-items-center justify-content-center">
         <%= link_to 'Scan QR code', '/associations/new', class: "btn btn-success center" %>
       </div>
+      
+      <div class="row align-items-center justify-content-center">
+        <%= select_tag name='event', options_from_collection_for_select(@events, "id", "name") , class: "btn btn-success center" %>
+      </div>
+
     </div>
   </div>
 </div>
